requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
flask-kafka==0.1.0
langchain==0.3.7
langchain-openai==0.2.6
openai==1.54.0
python-dotenv==1.0.1